Debugging 2nd scheduler -Passing load-60, load-avg

+++ src/threads/thread.h
@@ -96,7 +96,7 @@ struct thread {
     char name[16];                      /*!< Name (for debugging purposes). */
     uint8_t *stack;                     /*!< Saved stack pointer. */
     int nice;                           /*!< Niceness of the thread.*/
-    int64_t recent_cpu;                     /*!< Recent_cpu of the thread */
+    int32_t recent_cpu;                     /*!< Recent_cpu of the thread */
     int priority;                       /*!< Priority. */
     int donated_priority;
     struct list_elem allelem;           /*!< List element for all threads list. */
